Extract formatTime to utils for reusability
